refactor : Racing, Winner
